SRGAN/requirements.txt
matplotlib
pandas
numpy
math
torch
torchvision
pillow
tqdm
torchmetrics